+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### 新特性
 * 【core】         MapUtil增加toObjectArray方法
 * 【core】         URLUtil.normalize增加反斜杠处理（issue#IM8BI@Gitee）
+* 【core】         增加ClassUtil.getShortClassName（issue#IM8XM@Gitee）
 
 ### Bug修复